feat(dsg): Duplicate createDockerComposeDBFile with new name to replace it
