+++ packages/data-service-generator/src/server/docker-compose/docker-compose.dev.template.yml
